requirements.txt
discord.py==2.3.2
python-dotenv==1.0.1
requests==2.28.1
aiofiles==23.2.1
pyspellchecker==0.8.1
gitpython==3.1.42